[Feat] Comment 댓글 기능

+++ src/main/java/com/web/app/service/ReviewServiceImpl.java
@@ -1,17 +1,25 @@
 package com.web.app.service;
 
 import com.web.app.domain.board.Board;
+import com.web.app.domain.comment.Comment;
 import com.web.app.domain.review.Review;
+import com.web.app.dto.CommentResponseDTO;
 import com.web.app.dto.ReviewListDTO;
 import com.web.app.dto.ReviewRequestDTO;
 import com.web.app.dto.ReviewResponseDTO;
+import com.web.app.mediator.GetEmailFromJWT;
 import com.web.app.repository.BoardRepository;
+import com.web.app.repository.CommentRepository;
+import com.web.app.repository.LikesRepository;
 import com.web.app.repository.ReviewRepository;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
@@ -20,12 +28,23 @@ public class ReviewServiceImpl implements ReviewService {
 
     private final ReviewRepository reviewRepository;
     private final BoardRepository boardRepository;
+    private final CommentRepository commentRepository;
     private final ModelMapper modelMapper;
+    private final GetEmailFromJWT getEmailFromJWT;
+    private final LikesRepository likesRepository;
+
 
     @Override
-    public Long register(Review review, Long boardId) {
+    public Long register(HttpServletRequest request, Review review, Long boardId) {
 
         Board board = boardRepository.findById(boardId).orElseThrow();
+
+        String requestEmail = getEmailFromJWT.execute(request);
+
+        if (!board.getEmail().equals(requestEmail) || requestEmail.isBlank()) {
+            throw new RuntimeException("해당 요청은 게시글 작성자만 가능합니다.");
+        }
+
         review.setBoard(board);
         Review saved = reviewRepository.save(review);
 
@@ -33,11 +52,26 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
     @Override
-    public ReviewResponseDTO read(Long id) {
+    public ReviewResponseDTO read(Long id, HttpServletRequest request) {
 
         Review findReview = reviewRepository.findById(id).orElseThrow();
 
         List<Review> reviews = reviewRepository.findAllByBoardIsOrderByIdDesc(findReview.getBoard());
+
+        List<Comment> comments = commentRepository.findAllByBoardIsOrderByCreatedAtDesc(findReview.getBoard());
+
+        String requestEmail = getEmailFromJWT.execute(request);
+
+        List<Long> liked;
+        if (requestEmail.isBlank()) {
+            liked = new ArrayList<>();
+        } else {
+            liked = likesRepository.isLiked(findReview.getBoard().getId(), requestEmail);
+        }
+
+        List<CommentResponseDTO> commentListDTO = comments.stream()
+                .map(comment -> new CommentResponseDTO(comment, requestEmail))
+                .collect(Collectors.toList());
 
         List<ReviewListDTO> reviewListDTOS = reviews.stream()
                 .map(review -> modelMapper.map(review, ReviewListDTO.class))
@@ -45,13 +79,29 @@ public class ReviewServiceImpl implements ReviewService {
 
         ReviewResponseDTO responseDTO = findReview.toResponseDTO(reviewListDTOS);
 
+        if (!findReview.getBoard().getEmail().equals(requestEmail) || requestEmail.isBlank()) {
+            responseDTO.setMyBoard(false);
+        } else {
+            responseDTO.setMyBoard(true);
+        }
+
+        responseDTO.setCommentList(commentListDTO);
+        responseDTO.setLiked(!liked.isEmpty());
+        responseDTO.setLikeCount(findReview.getBoard().getLikeCount());
+
         return responseDTO;
     }
 
     @Override
-    public Review modify(ReviewRequestDTO reviewRequestDTO , Long id) {
+    public Review modify(HttpServletRequest request, ReviewRequestDTO reviewRequestDTO , Long id) {
 
         Review findOne = reviewRepository.findById(id).orElseThrow();
+
+        String requestEmail = getEmailFromJWT.execute(request);
+
+        if (!findOne.getBoard().getEmail().equals(requestEmail) || requestEmail.isBlank()) {
+            throw new RuntimeException("해당 요청은 게시글 작성자만 가능합니다.");
+        }
 
         findOne.change(reviewRequestDTO.getSubTitle(), reviewRequestDTO.getContent()); // dirty check ? => not
 
@@ -61,9 +111,16 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
     @Override
-    public void remove(Long id) {
-        // TODO: 삭제 실패 시 예외 처리 추가
-        reviewRepository.deleteById(id);
+    public void remove(HttpServletRequest request, Long id) {
 
+        Review review = reviewRepository.findById(id).orElseThrow();
+
+        String requestEmail = getEmailFromJWT.execute(request);
+
+        if (!review.getBoard().getEmail().equals(requestEmail) || requestEmail.isBlank()) {
+            throw new RuntimeException("해당 요청은 게시글 작성자만 가능합니다.");
+        }
+
+        reviewRepository.deleteById(id);
     }
 }
